normalement tout devrait etre safe now

+++ application/controllers/Process_secretariat.php
@@ -28,14 +28,17 @@ class Process_secretariat extends CI_Controller {
   }
 
   public function addUEtoParcours(){
-    //TODO verification des droit sur le liens UE - PArcours
     $this->load->model("administration_model",'adminMod');
     $ids = array();
-    foreach ($_GET['idUEs'] as $idUE) {
+    if($this->adminMod->isThisParcoursEditable($_GET['idParcours'])) {
 
-      if($this->adminMod->addUEtoParcours($_GET['idParcours'],$idUE)){
-        $ids[] = $idUE;
+      foreach ($_GET['idUEs'] as $idUE) {
+
+        if($this->adminMod->addUEtoParcours($_GET['idParcours'],$idUE)){
+          $ids[] = $idUE;
+        }
       }
+
     }
     header('Content-Type: application/json');
 
@@ -43,23 +46,24 @@ class Process_secretariat extends CI_Controller {
   }
 
   public function removeUEtoParcours(){
-    //TODO verification des droit sur le liens UE - PArcours
-
     $this->load->model("administration_model",'adminMod');
     $ids = array();
-    foreach ($_GET['idUEs'] as $idUE) {
-      if($this->adminMod->removeUEtoParcours($_GET['idParcours'],$idUE)){
-        $ids[] = $idUE;
+    if($this->adminMod->isThisParcoursEditable($_GET['idParcours'])) {
+
+      foreach ($_GET['idUEs'] as $idUE) {
+        if($this->adminMod->removeUEtoParcours($_GET['idParcours'],$idUE)){
+          $ids[] = $idUE;
+        }
       }
     }
     header('Content-Type: application/json');
 
     echo json_encode($ids);
+
   }
 
   public function addParcours(){
     $this->load->model("administration_model",'adminMod');
-
     if(isset($_POST['send']) && isset($_POST['year']) && isset($_POST['type'])){
       if(is_numeric($_POST['year']) && strlen($_POST['type']) == 2){
         if($this->adminMod->addParcours($_POST['year'],$_POST['type'])){
@@ -78,13 +82,15 @@ class Process_secretariat extends CI_Controller {
 
   public function deleteParcours(){
     $this->load->model("administration_model",'adminMod');
-    //TODO verification des droit sur le liens UE - PArcours
-    //TODO gerer le cas des semestre deja relier en plus du js I mean :)
     if(isset($_POST['parcours'])){
-      if($this->adminMod->deleteCascadeParcours($_POST['parcours'])){
-        $this->session->set_flashdata("notif", array("Parcours supprimé !"));
+      if($this->adminMod->isThisParcoursEditable($_POST['parcours'])) {
+        if($this->adminMod->deleteCascadeParcours($_POST['parcours'])){
+          $this->session->set_flashdata("notif", array("Parcours supprimé !"));
+        }else{
+          $this->session->set_flashdata("notif", array("Erreur de suppression bdd !"));
+        }
       }else{
-        $this->session->set_flashdata("notif", array("Erreur de suppression bdd !"));
+        $this->session->set_flashdata("notif", array("Ce parcours ne peut etre supprimé !"));
       }
     }else{
       $this->session->set_flashdata("notif", array("Données manquantes !"));
